Fix setting the soft due date on every comment and other ticket changes.
It turns out that ticket._old does not contain the complete old ticket, but only the old values of the fields that were changed. Therefore, I can not check if priority is in ticket._old to see if this is a new ticket instead of a modification.

The new check takes into account all possible reasons to (re-)set a soft-due-date: the priority changed, the soft due date changed, or no soft due date was set yet.